Attempted Task 4
I have made the total value of what the coins pertain to, to increment to that value in the total value text box. For example, when selecting 1 pence the increment in the total value text box will be 1pence. For 20 pence the total value will be 20pence, and so on for each of the coins.