Attempted task 6 and have assigned reset button with functions
Here I have placed an “if statement” into "cost per credit" textbox by using "!=" meaning it will do nothing until a string is placed into it and Convert.ToInt32  means to convert the (string)text to an integer. So the "number of credits" will increment by 1 as the value of "cost per credit" textbox has reached its current value of what I placed it to be. Furthermore, I have set functions in the "reset button" to change all the variables in the textbox's and appropriate labels to reset back to "0". Once the "reset button" has been clicked.